Added friendship constraints to code
If the csv included friendships, they will be added to the input and create constraints accordingly.

Note that the number of constraints greatly increases due to these friendships and that it will almost always yield fractional solutions for problems with friendships. Only the simplest of friendship requests might be fulfilled with single-team assignments.