Inputs/outputs with friends clearer
Made the input csv writer append the friends to the file (they were previously missed). The result also prints the number of friends, the minimum size, maximum size, and average sizes of friendships (excluding singleton friendships).

MainFriendship and Main are updated to print the csv results. MainFriendship provides rounding for the friendship problem.